requirements.txt
openpyxl~=3.1.2
pandas~=2.0.3
numpy~=1.24.4
matplotlib~=3.7.2
seaborn~=0.12.2
PyYAML~=6.0.1
plotly~=5.15.0
bokeh~=3.3.0